+++ src/widgets/playlist/index.ts
@@ -6,3 +6,4 @@ export { default as ControlBar } from './ControlBar'
 export { default as PlaylistHorizontal } from './PlaylistHorizontal'
 export { default as PlaylistLayout } from './PlaylistLayout'
 export { default as YoutubePlayer } from './YoutubePlayer'
+export { default as PlaylistInfo } from './PlaylistInfo'
